+++ src/_globaldocs.c
@@ -13,10 +13,17 @@
  *
  *****************************************************************************/
 
-PyDoc_STRVAR(PYIU_reduce_doc, "Return state information for pickling.");
+PyDoc_STRVAR(PYIU_reduce_doc,
+             "__reduce__(/)\n--\n\n"
+             "Return a `tuple` containing the state information for pickling.");
 
-PyDoc_STRVAR(PYIU_setstate_doc, "Set state information for unpickling.");
+PyDoc_STRVAR(PYIU_setstate_doc,
+             "__setstate__(state, /)\n--\n\n"
+             "Set state for unpickling. "
+             "The `state` argument must be `tuple`-like.");
 
 #if PY_MAJOR_VERSION > 3 || (PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION >= 4)
-PyDoc_STRVAR(PYIU_lenhint_doc, "Return a length hint if possible.");
+PyDoc_STRVAR(PYIU_lenhint_doc,
+             "__length_hint__(/)\n--\n\n"
+             "Return an *estimate* for the length of the iterator or zero.");
 #endif