+++ src/roundrobin.c
@@ -136,18 +136,29 @@ roundrobin_reduce(PyIUObject_Roundrobin *self)
 {
     PyObject *ittuple, *res;
 
+    /* The "next" method modifies the "iteratortuple" so when someone uses
+       "reduce" they would see how it changes. So we need to return a copy
+       of that tuple here.
+       An additional side-effect of "next" is that exhausted iterators are
+       removed from the tuple, decremented and finally replaced by NULL at the
+       end of the tuple. Python interpreters shouldn't see these NULLs
+       (otherwise that might segfault) so we can simply "copy by slicing"
+       when there are already exhausted iterators inside.
+       */
     if (PyTuple_GET_SIZE(self->iteratortuple) != self->numactive) {
         ittuple = PyTuple_GetSlice(self->iteratortuple, 0, self->numactive);
-        if (ittuple == NULL) {
-            return NULL;
-        }
     } else {
-        ittuple = self->iteratortuple;
-        Py_INCREF(ittuple);
+        ittuple = PYUI_TupleCopy(self->iteratortuple);
     }
+    /* The error handling for both branches. */
+    if (ittuple == NULL) {
+        return NULL;
+    }
+
     res = Py_BuildValue("OO(nn)", Py_TYPE(self),
                         ittuple,
-                        self->numactive, self->active);
+                        self->numactive,
+                        self->active);
     Py_DECREF(ittuple);
     return res;
 }
@@ -162,8 +173,70 @@ roundrobin_setstate(PyIUObject_Roundrobin *self,
 {
     Py_ssize_t numactive, active;
 
-    if (!PyArg_ParseTuple(state, "nn", &numactive, &active)) {
+    if (!PyTuple_Check(state)) {
+        PyErr_Format(PyExc_TypeError,
+                     "`%.200s.__setstate__` expected a `tuple`-like argument"
+                     ", got `%.200s` instead.",
+                     Py_TYPE(self)->tp_name, Py_TYPE(state)->tp_name);
         return NULL;
+    }
+
+    if (!PyArg_ParseTuple(state, "nn:roundrobin.__setstate__",
+                          &numactive, &active)) {
+        return NULL;
+    }
+
+    /* active and numactive must be greater than zero, otherwise the "next"
+       method could access out-of-bounds indices for the iteratortuple.
+       */
+    if (active < 0 || numactive < 0) {
+        PyErr_Format(PyExc_ValueError,
+                     "`%.200s.__setstate__` expected that the first (%zd) and "
+                     "second (%zd) argument in the `state` are not negative.",
+                     Py_TYPE(self)->tp_name, numactive, active);
+        return NULL;
+    }
+    /* If numactive is not zero than the active must be strictly smaller than
+       numactive, otherwise the next "next" call would access an out of bounds
+       index of the iteratortuple (or NULL). If "numactive" is zero then
+       "active" must be zero as well (in this case it must not be greater but
+       equal).
+       */
+    if (numactive != 0 && active >= numactive) {
+        PyErr_Format(PyExc_ValueError,
+                     "`%.200s.__setstate__` expected that the first (%zd) "
+                     "argument in the `state` is strictly greater than the "
+                     "second (%zd) argument, if the first argument isn't zero.",
+                     Py_TYPE(self)->tp_name, numactive, active);
+        return NULL;
+    } else if (numactive == 0 && active != 0) {
+        PyErr_Format(PyExc_ValueError,
+                     "`%.200s.__setstate__` expected that the second (%zd) "
+                     "argument in the `state` is zero if the first "
+                     "argument (%zd) argument is zero.",
+                     Py_TYPE(self)->tp_name, active, numactive);
+        return NULL;
+    }
+
+    /* The "numactive" argument must match the number of not-NULL arguments
+       in the iteratortuple. Luckily the NULLs are at the end of the
+       "iteratortuple".
+       */
+    if (1) {
+        Py_ssize_t tupsize = PyTuple_GET_SIZE(self->iteratortuple);
+        /* decrement the tuple size as long as the last item is NULL. */
+        while (tupsize > 0 &&
+               PyTuple_GET_ITEM(self->iteratortuple, tupsize - 1) == NULL) {
+            tupsize--;
+        }
+        if (numactive != tupsize) {
+            PyErr_Format(PyExc_ValueError,
+                         "`%.200s.__setstate__` expected that the first "
+                         "argument in the `state` (%zd) is equal to the number "
+                         "of not exhausted iterators (%zd) in the instance.",
+                         Py_TYPE(self)->tp_name, numactive, tupsize);
+            return NULL;
+        }
     }
 
     self->numactive = numactive;
@@ -198,8 +271,8 @@ static PyMethodDef roundrobin_methods[] = {
 #if PY_MAJOR_VERSION > 3 || (PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION >= 4)
     {"__length_hint__", (PyCFunction)roundrobin_lengthhint, METH_NOARGS, PYIU_lenhint_doc},
 #endif
-    {"__reduce__", (PyCFunction)roundrobin_reduce, METH_NOARGS, PYIU_reduce_doc},
-    {"__setstate__", (PyCFunction)roundrobin_setstate, METH_O, PYIU_setstate_doc},
+    {"__reduce__",   (PyCFunction)roundrobin_reduce,   METH_NOARGS, PYIU_reduce_doc},
+    {"__setstate__", (PyCFunction)roundrobin_setstate, METH_O,      PYIU_setstate_doc},
     {NULL, NULL}
 };