+++ src/sideeffect.c
@@ -180,7 +180,10 @@ sideeffects_next(PyIUObject_Sideeffects *self)
                the "funcargs" and nobody except the instance and the funcargs
                owns the "collected". This can be up to 40-50% faster for small
                "times" values. Even for relativly bigger ones this is still
-               10% faster. */
+               10% faster.
+               To avoid needing to decrement the values in the tuple while
+               iterating these are simply set to NULL.
+               */
             if (Py_REFCNT(self->funcargs) == 1 &&
                     Py_REFCNT(self->collected) <= 2) {
                 for (i=0 ; i < self->times ; i++) {
@@ -212,21 +215,50 @@ Fail:
 static PyObject *
 sideeffects_reduce(PyIUObject_Sideeffects *self)
 {
-    Py_ssize_t i;
-    PyObject *temp;
-    /* Need to set all unset elements to None otherwise not possible to
-       pickle the collected tuple. */
-    for (i=0 ; i < self->times ; i++) {
-        temp = PyTuple_GET_ITEM(self->collected, i);
-        if (temp == NULL) {
+    PyObject *collected;
+    PyObject *res;
+
+    /* There are several issues that prevent from simply wrapping the
+       attributes.
+       */
+    if (self->collected == NULL) {
+        /* When "collected" is NULL we wrap it as None, and no further
+           processing is needed.
+           */
+        Py_INCREF(Py_None);
+        collected = Py_None;
+    } else {
+        /* When we have "collected" then it's a tuple that may contain NULLs.
+           The Python interpreter does not like NULLs so these must be
+           replaced by some fillvalue (in this case None). However we modify
+           the tuple inside the "next" method so if someone called "reduce"
+           that person could **see** the "collected" tuple change. That must
+           be avoided so we MUST return a copy of the "collected" tuple.
+           */
+        Py_ssize_t i;
+        Py_ssize_t collected_size = PyTuple_GET_SIZE(self->collected);
+        collected = PyTuple_New(collected_size);
+        if (collected == NULL) {
+            return NULL;
+        }
+        for (i = 0 ; i < collected_size ; i++) {
+            PyObject *tmp = PyTuple_GET_ITEM(self->collected, i);
+            if (tmp == NULL) {
+                tmp = Py_None;
+            }
             Py_INCREF(Py_None);
-            PyTuple_SET_ITEM(self->collected, i, Py_None);
+            PyTuple_SET_ITEM(collected, i, tmp);
         }
     }
-    return Py_BuildValue("O(OOn)(nO)", Py_TYPE(self),
-                         self->iterator, self->func, self->times,
-                         self->count,
-                         self->collected ? self->collected : Py_None);
+
+    res = Py_BuildValue("O(OOn)(nO)", Py_TYPE(self),
+                        self->iterator,
+                        self->func,
+                        self->times,
+                        self->count,
+                        collected);
+    Py_DECREF(collected);
+    return res;
 }
 
 /******************************************************************************
@@ -239,17 +271,114 @@ sideeffects_setstate(PyIUObject_Sideeffects *self,
 {
     Py_ssize_t count;
     PyObject *collected;
-    if (!PyArg_ParseTuple(state, "nO", &count, &collected)) {
+    PyObject *newcollected = NULL;
+    Py_ssize_t collected_size;
+
+    if (!PyTuple_Check(state)) {
+        PyErr_Format(PyExc_TypeError,
+                     "`%.200s.__setstate__` expected a `tuple`-like argument"
+                     ", got `%.200s` instead.",
+                     Py_TYPE(self)->tp_name, Py_TYPE(state)->tp_name);
         return NULL;
+    }
+
+    if (!PyArg_ParseTuple(state, "nO:sideeffects.__setstate__",
+                          &count, &collected)) {
+        return NULL;
+    }
+
+    /* The "collected" argument should be a tuple (because we use
+       PyTuple_GET_ITEM and PyTuple_SET_ITEM and thus would risk segmentation
+       faults if we don't check that it's a tuple) or None.
+       */
+    if (PyTuple_CheckExact(collected)) {
+        /* The class itself has a "times" attribute, if that attribute is zero
+           we do not need a "collected" tuple, it should have been "None".
+           */
+        if (self->times == 0) {
+            PyErr_Format(PyExc_TypeError,
+                         "`%.200s.__setstate__` expected `None` as second "
+                         "argument in the `state` when `self->times == 0`, "
+                         "got %R.",
+                         Py_TYPE(self)->tp_name, collected);
+            return NULL;
+        }
+        /* The "count" must not be negative or bigger/equal to the size of
+           the "collected" tuple. Otherwise we would access indices that are
+           out of bounds for the tuple in "next".
+           */
+        collected_size = PyTuple_GET_SIZE(collected);
+        if (count < 0 || count >= collected_size) {
+            PyErr_Format(PyExc_ValueError,
+                         "`%.200s.__setstate__` expected that the first "
+                         "argument in the `state` (%zd) is not negative and "
+                         "smaller than the length of the second argument "
+                         "(%zd).",
+                         Py_TYPE(self)->tp_name, count, collected_size);
+            return NULL;
+        }
+        /* The length of the "collected" tuple must also be equal to the
+           "self->times" attribute.
+           */
+        if (self->times != collected_size) {
+            PyErr_Format(PyExc_ValueError,
+                         "`%.200s.__setstate__` expected that the second "
+                         "argument in the `state` has a length (%zd) "
+                         "equal to the `self->times` (%zd) attribute.",
+                         Py_TYPE(self)->tp_name, collected_size, self->times);
+            return NULL;
+        }
+    } else if (collected == Py_None) {
+        /* We only expect None if self->times and count is zero. */
+        if (count != 0 || self->times != 0) {
+            PyErr_Format(PyExc_TypeError,
+                         "`%.200s.__setstate__` expected a `tuple` as second "
+                         "argument in the `state` when `self->times != 0` or "
+                         "the first argument in the `state` is not zero, "
+                         "got None",
+                         Py_TYPE(self)->tp_name);
+            return NULL;
+        }
+    } else {
+        PyErr_Format(PyExc_TypeError,
+                     "`%.200s.__setstate__` expected a `tuple` or `None` as "
+                     "second argument in the `state`, got %.200s",
+                     Py_TYPE(self)->tp_name, Py_TYPE(collected)->tp_name);
+        return NULL;
+    }
+
+    /* In any case we need to process the "collected" value. In case it is
+       "None" we simply set it to NULL. However if it's not None then it's
+       a tuple. We process the tuple in the "next" function but it's possible
+       that someone still holds a reference to the tuple he passed in. So to
+       make sure that we don't mutate tuples that are in use elsewhere we
+       create a new tuple here. That also has the additional advantage that
+       we can leave the values with index below "count" as NULL. The "next"
+       method assumes that it doesn't have to decrement items that it sets so
+       this makes sure we don't create a memory leak there.
+       */
+    if (collected != Py_None) {
+        Py_ssize_t i;
+        newcollected = PyTuple_New(collected_size);
+        if (newcollected == NULL) {
+            return NULL;
+        }
+        for (i=0 ; i < count ; i++) {
+            PyObject *tmp = PyTuple_GET_ITEM(collected, i);
+            Py_INCREF(tmp);
+            PyTuple_SET_ITEM(newcollected, i, tmp);
+        }
+    } else {
+        newcollected = NULL;
     }
 
     self->count = count;
 
-    if (collected != Py_None) {
-        Py_CLEAR(self->collected);
-        Py_INCREF(collected);
-        self->collected = collected;
-    }
+    /* We already created a new tuple for "collected" or it's None so no need
+       to increment the reference count here.
+       */
+    Py_CLEAR(self->collected);
+    self->collected = newcollected;
 
     Py_RETURN_NONE;
 }
@@ -274,8 +403,8 @@ static PyMethodDef sideeffects_methods[] = {
 #if PY_MAJOR_VERSION > 3 || (PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION >= 4)
     {"__length_hint__", (PyCFunction)sideeffects_lengthhint, METH_NOARGS, PYIU_lenhint_doc},
 #endif
-    {"__reduce__", (PyCFunction)sideeffects_reduce, METH_NOARGS, PYIU_reduce_doc},
-    {"__setstate__", (PyCFunction)sideeffects_setstate, METH_O, PYIU_setstate_doc},
+    {"__reduce__",   (PyCFunction)sideeffects_reduce,   METH_NOARGS, PYIU_reduce_doc},
+    {"__setstate__", (PyCFunction)sideeffects_setstate, METH_O,      PYIU_setstate_doc},
     {NULL, NULL}
 };